Prevent unauthorized user to access protected url (create, edit, destroy).

+++ app/Http/Controllers/BonusController.php
@@ -11,8 +11,8 @@ use App\Models\Bonus;
 use App\Models\BonusDetail;
 use App\Models\Employee;
 use Illuminate\Database\Eloquent\Builder;
-use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class BonusController extends Controller
 {
@@ -79,6 +79,13 @@ class BonusController extends Controller
 
     public function edit(string $id)
     {
+        $response = Gate::inspect('update', Bonus::class);
+        $message = '';
+        if ($response->denied()) {
+            $message = $response->message();
+            return inertia('Error/Unauthorized', ['message' => $message]);
+        }
+
         $bonus = Bonus::find($id);
         $bonus_detail = BonusDetail::where('bonus_id', $id)->get();
         $employees = EmployeeResource::collection(Employee::all());
@@ -92,6 +99,13 @@ class BonusController extends Controller
 
     public function update(BonusUpdateRequest $request, string $id)
     {
+        $response = Gate::inspect('update', Bonus::class);
+        $message = '';
+        if ($response->denied()) {
+            $message = $response->message();
+            return inertia('Error/Unauthorized', ['message' => $message]);
+        }
+
         $bonus = Bonus::find($id);
         $bonus->update($request->validated());
 
@@ -114,6 +128,13 @@ class BonusController extends Controller
 
     public function destroy(string $id)
     {
+        $response = Gate::inspect('delete', Bonus::class);
+        $message = '';
+        if ($response->denied()) {
+            $message = $response->message();
+            return inertia('Error/Unauthorized', ['message' => $message]);
+        }
+
         $bonus_detail = BonusDetail::where('bonus_id', $id)->get();
         foreach ($bonus_detail as $line) {
             $line->delete();
